Moved utils.py up a directory.
